% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dc.R
\name{sp_dc}
\alias{sp_dc}
\title{Title Performs sparse-projection method under distributed setting}
\usage{
sp_dc(XtX, Xty, R = 2000, p)
}
\arguments{
\item{XtX}{Sum of t(X)\%*\%X from different datasets/machines}

\item{Xty}{Sum of t(X)\%*\%y from different datasets/machines}

\item{R}{Number of posterior samples}

\item{p}{Data dimensionality}
}
\value{
A list containing the following components:
\describe{
\item{post.est}{The posterior estimator of the regression coefficients.}
\item{post.dist}{The p*R posterior samples.}
\item{select.model}{The median probability model selected by the sparse-projection method.}
\item{confint}{The debiased-projection confidence regions.}
}
}
\description{
Title Performs sparse-projection method under distributed setting
}
