% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseProj.R
\name{sparseProj}
\alias{sparseProj}
\title{Title}
\usage{
sparseProj(X, y, Z, R = 2000)
}
\arguments{
\item{X}{n*p dimensional design matrix.}

\item{y}{n*1 dimensional response variable.}

\item{Z}{Intermediate result to be obtained from ZnZ.}

\item{R}{Number of posterior samples.}
}
\value{
A list containing the following components:
\describe{
\item{post.est}{The posterior estimator of the regression coefficients.}
\item{post.dist}{The p*R posterior samples.}
\item{select.model}{The median probability model selected by the sparse-projection method.}
\item{confint}{The debiased-projection confidence regions.}
}
}
\description{
Title
}
\examples{
X <- matrix(rnorm(50*100), nrow = 50, ncol = 100)
b <- c(rep(2,10),rep(0,100-10))
y <- X\%*\%b + rnorm(50)
Z <- ZnZ(X,y)
fit <- sparseProj(X,y,Z)

}
\author{
Samhita Pal
}
