% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZnZ.R
\name{ZnZ}
\alias{ZnZ}
\title{Title}
\usage{
ZnZ(X, y)
}
\arguments{
\item{X}{n*p design matrix.}

\item{y}{n*1 response variable.}
}
\value{
The Z matrix to be used in sparseProj.
}
\description{
Title
}
\examples{
Z <- ZnZ(X,y)

}
\author{
Samhita Pal
}
